+++ packages/client/src/auth.ts
@@ -1,0 +1,9 @@
+export interface MessagingClientAuth {
+  clientToken?: string
+  webhookToken?: string
+}
+
+export interface MessagingClientCredentials {
+  clientId: string
+  clientToken: string
+}
