+++ packages/client/src/index.ts
@@ -1,4 +1,8 @@
 export * from '@botpress/messaging-base'
+export * from './auth'
 export * from './channel'
 export * from './client'
 export * from './errors'
+export * from './events'
+export * from './options'
+export * from './schema'
